+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   * Change: `CallController#join` now blocks until a corresponding unjoined event is received
   * Feature: Added `Call#unjoin`
   * Feature: `CallControll#join` can be made non-blocking by passing `:async => true`
+  * Feature: `CallController#ask` now supports `:terminator` and `:limit` options to add early termination by terminator digit or digit limit conditions
 
 # 2.0.0.beta1 - 2012-03-07
   * Bugfix: #speak now correctly casts the argument to string if it is not SSML